+++ metfilelib/util/file_reader.py
@@ -1,0 +1,65 @@
+# Python 3 is coming to town
+from __future__ import print_function, unicode_literals
+from __future__ import absolute_import, division
+
+import os
+import sys
+
+
+class FileReader(object):
+    """A simple file reader for line-based files. It provides a line
+    lookahead and line numbering, and can log error messages. Error
+    messages are sent to stderr by default, but it is possible to
+    provide another error message handler (any object with a write()
+    method).
+
+    Files are opened with universal line endings (\n, \r\n or
+    \n\r)."""
+
+    def __init__(self, file_path, error_message_handler=sys.stderr):
+        """File_path is a full path to a file. It will be opened immediately,
+        so if any exceptions occur, they will probably happen here.
+
+        This immediately reads the first line of the file.
+
+        error_message_handler is any object with a write() method. It
+        will receive error messages of the form "filename 1: error
+        message\n", where "filename" is the basename of the file_path
+        and "1" is the line number.
+        """
+
+        self._open_file = open(file_path, 'rU')
+        self.filename = os.path.basename(file_path)
+        self._error_message_handler = error_message_handler
+        self.line_number = 0
+        self.current_line = None
+        self.success = True
+
+        # Read the first line
+        self.next()
+
+    def next(self):
+        """Read the next line, if we are not already at EOF. Updates
+        self.current_line and self.line_number. May result in a change
+        in self.eof."""
+        if not self.eof:
+            self.current_line = self._open_file.readline()
+            self.line_number += 1
+
+    @property
+    def eof(self):
+        """Return True if we are at end of file."""
+        return self.current_line == ''
+
+    def make_error(self, error_message):
+        """Format an error message and send it to the error message handler.
+
+        After this function is called at least once, self.success is
+        False."""
+        error = "{filename} {line}: {error}\n".format(
+            filename=self.filename,
+            line=self.line_number,
+            error=error_message)
+
+        self._error_message_handler.write(error)
+        self.success = False