+++ CHANGES.rst
@@ -6,3 +6,7 @@ Changelog of metfilelib
 ----------------
 
 - Initial project structure created with nensskel 1.31.dev0.
+
+- Added a util.file_reader.FileReader class that will probably be used
+  by the site, or some object with the same API. The metfile parser
+  should assume it gets instances of this class.